third_party/WebKit/LayoutTests/bluetooth/characteristic/service-same-from-2-characteristics.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = "Same parent service returned from multiple characteristics.";

bluetooth_test(() => getHealthThermometerDevice()
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => {
      return Promise.all([
          service.getCharacteristic('measurement_interval'),
          service.getCharacteristic('temperature_measurement')
      ]);
    })
    .then(characteristics => {
      assert_equals(characteristics[0].service, characteristics[1].service);
}), test_desc);
</script>
